코드 정리: idea 스크립트 problems 반영

+++ tomcat_172.28.200.101/get.sh
@@ -1,16 +1,16 @@
 #!/bin/bash
 
-LIST=($(cat <<- EOF
+mapfile -t LIST < <(cat <<- EOF
 
 172.28.200.101:tomcat
 
 EOF
-))
+)
 
 
-read -r rows <<< ${LIST[*]}
+read -r rows <<< "${LIST[@]}"
 idx=0
-for item in ${LIST[*]}; do
+for item in "${LIST[@]}"; do
   ((idx++))
   if [[ "${item}" == "#"* ]]; then
     printf "\e[0;31m ## ${idx}) ${item} (SKIP) \e[0m%s" $'\n'
@@ -24,30 +24,30 @@ for item in ${LIST[*]}; do
   outDir="${hostnm}_${ip}"
   printf "\e[0;31m ## ${idx}) ${ip} ${hostnm} \e[0m%s" $'\n'
   
-  if [ ! -e ${outDir} ]; then
-    mkdir ${outDir}
+  if [ ! -e "${outDir}" ]; then
+    mkdir "${outDir}"
   fi
   
-  ssh root@${ip} "sh /root/sh/host-backup.sh"
+  ssh "root@${ip}" "sh /root/sh/host-backup.sh"
   
   printf " transering ... "
-  ssh root@${ip} "tar cfz - /root/${gzFile} 2> /dev/null" \
-    | tar xvfz - --strip-components=1 -C ${outDir} 2>&1 > /dev/null
+  ssh "root@${ip}" "tar cfz - /root/${gzFile} 2> /dev/null" \
+    | tar xvfz - --strip-components=1 -C "${outDir}" > /dev/null 2>&1
   printf "done%s" $'\n'
   
-  if [ ${gzFile} != "" ]; then
+  if [ "${gzFile}" != "" ]; then
     printf " deleting remote ${gzFile} ... "
-    ssh root@${ip} "rm -rf /root/${gzFile}"
+    ssh "root@${ip}" "rm -rf /root/${gzFile}"
     printf "done%s" $'\n'
   fi
   
   printf " extracting ${gzFile} to ./${outDir}/ ... "
-  tar xfz ${outDir}/${gzFile} -C ./${outDir}/ 2>&1 > /dev/null
+  tar xfz "${outDir}/${gzFile}" -C ./"${outDir}"/ > /dev/null 2>&1
   printf "done%s" $'\n'
   
-  if [ -e ${outDir}/${gzFile} ]; then
+  if [ -e "${outDir}/${gzFile}" ]; then
     printf " deleting local ${gzFile} ... "
-    rm -rf ${outDir}/${gzFile}
+    rm -rf "${outDir:?}/${gzFile:?}"
     printf "done%s" $'\n'
   fi
 done
